more integration, moving methods onto the Resource classes.
